Add Instant Run support (#78)
* Refactor PackageTree to de-duplicate added refs
* Add support for Instant Run builds
In preparation for Instant Run support, we need to make PackageTree
remember which methods and fields have been added.  Instant Run will
include many different .dex files, each with overlapping method
references which need to be handled if we want counts to make sense.

* Add support for Instant Run builds

Building on the last commit, we now search for 'instant-run.zip' within
APK files.  If it is found, we unzip it and extract all of the
incremental dex files it contains.  These, along with the expected
'classes(\d?).dex' files, are fed into PackageTree which now knows how
to deduplicate method refs.

This appears to work - as best as I can tell (via shell-script hackery
and experimentation), this accurately counts app code plus the added
Instant Run instrumentation.  It's not possible (rather, not at all
feasible) to distinguish between user code and instrumentation.